requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database dependencies
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.7
numpy==1.24.3
torch>=1.13.0

# Data collection APIs
yfinance==0.2.28
tiingo==0.14.0
finnhub-python==2.4.20
requests==2.31.0
websockets==12.0
aiohttp==3.9.1

# ML and NLP
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing
beautifulsoup4==4.12.2
feedparser==6.0.10
websockets==12.0
python-dateutil==2.8.2
pytz==2023.3

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Task scheduling
celery==5.3.4
redis==5.0.1
schedule==1.2.0

# HTTP client
httpx==0.25.2

# Data validation
marshmallow==3.20.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional ML serving (commented out - requires cmake)
# onnx==1.15.0
# onnxruntime==1.16.3